Device can now output all its values at once. Fixed bug where the device was showing output on lcd when starting when there was no output

+++ Firmware/ADC.c
@@ -1,7 +1,7 @@
 #include "ADC.h"
 
 // define the raw ADC read variables
-uint16_t ADC_reading = 0;
+int ADC_reading = 0;
 unsigned char ADC_status = 0;
 
 // Setup for the ADC channels
